# GC reachability traversal in revorder