--- a/src/irmin-pack/unix/gc_args.ml
+++ b/src/irmin-pack/unix/gc_args.ml
@@ -58,6 +58,7 @@ module type S = sig
     val unsafe_find :
       check_integrity:bool -> [< read ] t -> key -> Node_value.t option
 
+    val key_of_offset : [< read ] t -> int63 -> key
     val purge_lru : 'a t -> unit
   end