--- a/src/irmin-pack/unix/gc_args.ml
+++ b/src/irmin-pack/unix/gc_args.ml
@@ -59,6 +59,7 @@ module type S = sig
       check_integrity:bool -> [< read ] t -> key -> Node_value.t option
 
     val key_of_offset : [< read ] t -> int63 -> key
+    val unsafe_find_no_prefetch : 'a t -> key -> Node_value.t option
     val purge_lru : 'a t -> unit
   end