--- a/src/irmin-pack/unix/inode_intf.ml
+++ b/src/irmin-pack/unix/inode_intf.ml
@@ -57,6 +57,7 @@ module type Persistent = sig
   val to_snapshot : Raw.t -> Snapshot.inode
   val of_snapshot : 'a t -> index:(hash -> key) -> Snapshot.inode -> value
   val purge_lru : 'a t -> unit
+  val key_of_offset : 'a t -> int63 -> key
 end
 
 module type Sigs = sig