I've been looking at the GC reachability traversal. I still have some ideas but I would like to avoid going down too far the wrong path because I overlooked something... The quoted numbers are from the Tezos replay trace, running the GC pretty aggressively (so real life might behave differently? by how much?)

- The GC traversal of the tree uses a hashtable to avoid revisiting the same objects again (because it's actually a DAG). However it only checks the hashtable **after** re-inserting the `(offset,length)` into the reachable file: The reachable file ends up with 28m pairs (~440Mb) that gets sorted/deduplicated in ~15s. On the Tezos replay trace, I'm seeing a lot of duplication as there are actually 14m real objects (6m contents + 8m nodes, with indistinct repetitions for both kinds). By checking the hashtable first before inserting, the `store.X.reachable` file nearly halves (~230Mb) and its sorting time too.

- However, the hashtable now contains 14m entries rather than 8m (as only the nodes where kept before.) https://github.com/mirage/irmin/issues/2039 mentions that the GC reachability could traverse the disk in-order. By using a priority queue on the decreasing offsets, we can clear the hashtable from old entries that will not be visited again... But it's not very clear that the priority queue won't explose in size: I'm seeing it stabilize at max ~4m elements. I'm guessing that the worst case would be `max |contents| |nodes|` with an unlikely use of irmin that would lay out the tree in bfs order on the disk before GC.

- At this point, the `store.X.reachable` gets almost produced in reverse sorted order thanks to the priority queue. Only the GC commit's parents are inserted at the wrong time, but this is easily fixed by including them in the in-order traversal. We can then skip sorting `store.X.reachable` and just reverse it to compute the `store.X.mapping` file directly. This removes the need for the large intermediate `store.X.sorted` file.
  (Sorting with the priority queue during traversal is a in theory a bit more efficient than a general purpose sort, as it can exploit the backward edges characteristic of the irmin store) 

- Finally, the mapping file collapses consecutive `(offset,length)` ranges. This can now be done much earlier when producing the `store.X.reachable`, such that this file drops to the final size of `store.X.mapping` (which is much smaller.)

I believe that this PR should already be a bit faster (it at least saves a few Mb on disk), although the following "GC total duration" from the replay trace should be taken with a grain of salt as I'm seeing a lot of variations across multiple runs on my computer:

```
Before:  64.9s  130.7s  140.5s  151.5s  154.1s  151.6s  154.9s
 After:  34.8s   87.6s  104.5s  106.7s  110.1s  124.3s  118.5s
```

Some pending questions/notes:
- How bad is the max memory consumption with the priority queue? (I'm hoping it should be about the same as before)
- The in-disk-order traversal is not quite there yet, as the node's `decode_bin` does too much for our purpose (including arbitrary disk reads)
- I was lazy and duplicated some routines in `Mapping_file` (which I can clean up), but I'm not sure if we should keep the original dead(ly interesting) code? (it would still be available in the git history if needed)
- The `store.X.reachable` could be reversed in place to save a few Mb... Do we know how large it gets in practice?