@art-w 2022-09-21
> Do you know how big the sorted file is? You say it is large, so it's very nice that this PR eliminates this disk usage!

Oh sorry for the confusion, it's only "large" when compared to the final `store.X.mapping` file. The `store.X.sorted` size was ~440Mb at the start (+ same size for `store.X.reachable` which needs to exist at the same time), then half with deduplication, then 0 since we don't need it anymore. So this saves less than 1Gb in the end for the Tezos replay trace.

I went ahead to reverse the mapping file in place, not because it saves the last few Mb, but because it's one less temporary file that we need to clean up in case of failure (... and the code is a bit shorter/simpler from my perspective ^^')