@irmaTS 2022-09-21
> 


> I went ahead to reverse the mapping file in place, not because it saves the last few Mb, but because it's one less temporary file that we need to clean up in case of failure (... and the code is a bit shorter/simpler from my perspective ^^')

If this means that the mapping file used by `main` is modified by the GC worker, then this breaks the RW consistency.