@art-w 2022-09-21
It would break everything yes! But it is the new mapping file before it becomes visible and used by anyone :)