@Ngoguey42 2022-09-26
I would like to benchmark this PR before merging. I'm still working on the stats