@icristescu 2022-11-09
> Here is the GC benchmark that compares main with this PR. I have not looked in detail
> [pp-gc.txt](https://github.com/mirage/irmin/files/9931503/pp-gc.txt)

Thanks, I had a quick look - these metrics are super nice to have btw, great work!

The worker total time is a bit down, but the maxrss is a bit up:
```
                       | with-gc_a  |    with-gc_b    |    pr-2085_a    |    pr-2085_b
total duration (wall)  |    5min42s |    5min45s 101% |    5min31s  97% |    5min32s  97%

initial maxrss (bytes) |   75.308 M |   75.312 M 100% |   75.176 M 100% |   77.160 M 102%
final maxrss (bytes)   | 1129.136 M | 1129.312 M 100% | 1274.100 M 113% | 1274.180 M 113%
```
I think we should go forward with this PR, the code is cleaner and the maxrss overhead is not so big, IMHO. 

The diffs in time and maxrss come from where we expected it:
```
-- Worker's timings --
                               |           |      wall      | share |      sys       |      user      | %cpu | %sys | %user
mapping: of reachable          | with-gc_a | 13.306  s      |  3.9% |  0.503  s      | 12.311  s      |  96% | 3.8% |   93%
                               | with-gc_b | 13.300  s 100% |  3.9% |  0.769  s 153% | 12.022  s  98% |  96% | 5.8% |   90%
                               | pr-2085_a |  0.162  s 1.2% |  .05% |  0.027  s 5.3% |  0.036  s .29% |  39% |  17% |   22%
                               | pr-2085_b |  0.161  s 1.2% |  .05% |  0.029  s 5.7% |  0.028  s .22% |  35% |  18% |   17%
-- Worker's rusage stats per step --
                               |           |     maxrss     |    minflt    |    majflt    |    inblock     |    oublock     |    nvcsw     |   nivcsw
mapping: objects to reachable  | with-gc_a |   729_416      | 153_012      |       0      | 6_216_864      |   933_640      | 506_907      |  2_594     
                               | with-gc_b |   729_392 100% | 153_011 100% |       0  --  | 6_219_336 100% |   933_640 100% | 507_179 100% |  2_644 102%
                               | pr-2085_a | 1_131_032 155% | 254_503 166% |       0  --  | 7_081_456 114% |    82_072 8.8% | 475_903  94% |  2_686 104%
                               | pr-2085_b | 1_131_180 155% | 254_506 166% |       0  --  | 7_084_568 114% |    82_056 8.8% | 476_548  94% |  2_760 106%
mapping: of reachable          | with-gc_a | 1_129_136      | 127_831      | 118_376      |   406_744      | 1_016_928      |   1_093      |    427     
                               | with-gc_b | 1_129_312 100% | 126_266  99% | 118_368 100% |   404_752 100% | 1_016_928 100% |   1_042  95% |    426 100%
                               | pr-2085_a | 1_172_460 104% |     690 .54% |  10_355 8.7% |    13_672 3.4% |    72_632 7.1% |     226  21% |      5 1.2%
                               | pr-2085_b | 1_172_568 104% |     693 .54% |  10_363 8.8% |    15_632 3.8% |    70_680 7.0% |      95 8.7% |      8 1.9%

``` 

- this PR does not track all the stats as main? as there are some stats `n/a` or set to 0
- is it normal to have 0 in these columns?
```
appended_hashes        |          0 |          0  --  |          0  --  |          0  -- 
appended_offsets       |          0 |          0  --  |          0  --  |          0  -- 
total                  |  9_012_474 |  9_012_474 100% |  9_012_474 100% |  9_012_474 100%
```