@Ngoguey42 2022-11-09
> this PR does not track all the stats as main? as there are some stats n/a or set to 0

The stats only track give informations on main in `Main thread during GC` and `Main's timings`. The rest concerns the worker

> is it normal to have 0 in these columns?

Yes this is normal. The worker doesn't encore anything.