@art-w 2022-12-07
Rebased and added two commits to make this PR work:

- [irmin-pack: gc reachability, optimize memory by keeping only the key](https://github.com/mirage/irmin/pull/2085/commits/601d4438668e5df64f57acf456f052a202de8c6d)
  We can save quite a lot of memory during the GC traversal if we keep only the object offsets that have yet to be visited, and don't carry their hash and length (since we can recover that information from the offset)
- [irmin-pack: gc reachability, do not prefetch children hash/length](https://github.com/mirage/irmin/pull/2085/commits/55ab9b023138061bf8a3cdb741726c55a19993ee)
  This is a bit dirty, but it makes the GC traversal a lot faster: since we don't care anymore about the hash/length of a node children, we can avoid doing random reads on the disk by not prefetching those informations. I didn't find a nice way to do this, so I introduced a new unsafe `Pack_key.Offset` constructor that should never appear outside of the GC traversal.

I've removed a bunch of optimizations that didn't have a big impact, the latest benchmarks look like: (orange is still running to check what happens with more GC iterations, I'll update later)

![tezos_trio](https://user-images.githubusercontent.com/4807590/206211881-1166e5f8-ec50-49f8-a59b-57d429b7e91a.png)