@metanivek 2022-12-09
@art-w I know it's a tad annoying but would you mind applying the parts of [`ce23879` (#2085)](https://github.com/mirage/irmin/pull/2085/commits/ce23879d304efd93ec8f504c8b9996d268b47f82) to their appropriate commits that you're `fixup`ing? (any changes worthy of a new commit(s) should of course be self-standing)